Massive redesign (yes! that early in the game!)
This version has a serializer class and a deserializer class. I use streams
(ostream in the serializer and istream in the deserializer, but really it
can be any type T).

The serializer takes a writable stream. It will write bits and pieces to
it as the serialization happens. The recursivity doesn't use a size anymore.
The first version required to generate a sub-buffer in a vector, then copy
that vector in the previous level. This meant a lot of copies and a lot
of allocations/re-allocations. The stream such as the std::stringstream
will not re-allocate the same buffer over and over. Instead, it will allocate
new buffers to enlarge the file as required.

The deserializer is a separate object that you do not need to derive from
(and pollute your members). The class has a read_data() which returns the
content of the next field. That way, it can be saved directly in the final
variable buffer instead of some intermediary twice.