Updated to mention the new home: snapdev.
Since I changed it to just a single header file, there was no need for a library. So I could move this "library" (one header) to the snapdev project. The very first version included a .cpp, but that was removed. Also there is not logging so no need to depend on snaplogger.